// src/app/interface.ts
export interface carts {
    amount: number
    prod: products
}
export interface products {
    prod: any
    id: number
    title: string
    price: number
    description: string
    category: string
    image: string
    rating: Rating
}
export interface Rating {
    rate: number
    count: number
}